# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/user/admin": {
            "post": {
                "description": "Sign-up admin to system with email and password",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "SignUpAdmin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username user",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email user",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password user",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "User profile image",
                        "name": "files",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response\" example({\"message\":\"successful\",\"user_id\":\"uuid-123\",\"username\":\"john_doe\"})",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid email format, duplicate username, or duplicate email",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Password hashing error",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user/info": {
            "post": {
                "description": "create user info data",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "CreateUserInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username user",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "age user",
                        "name": "age",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"male\"",
                        "description": "gender user",
                        "name": "gender",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "default": 1.8,
                        "description": "height user",
                        "name": "height",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "default": 80,
                        "description": "weight user",
                        "name": "weight",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "default": 80,
                        "description": "target weight user",
                        "name": "target_weight",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"active\"",
                        "description": "active level user",
                        "name": "active_level",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response\" example({\"message\":\"successful\",\"user_id\":\"uuid-123\",\"username\":\"john_doe\"})",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid email format, duplicate username, or duplicate email",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Password hashing error",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user/list": {
            "get": {
                "description": "Get list users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "FetchAllUsers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "example: john doe",
                        "name": "search_word",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "example: 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "example: 10",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user/refresh-passport": {
            "get": {
                "description": "Refresh user passport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "RefreshUserPassport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "refresh_token",
                        "name": "refresh_token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user/sign-in": {
            "post": {
                "description": "Sign-in to system with email and password",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "SignIn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email user",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password user",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "password is invalid",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user/sign-up": {
            "post": {
                "description": "Sign-up to system with email and password",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "SignUp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username user",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "email user",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password user",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "user profile image",
                        "name": "files",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response\" example({\"message\":\"successful\",\"user_id\":\"uuid-123\",\"username\":\"john_doe\"})",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid email format, duplicate username, or duplicate email",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Password hashing error",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user/{user_id}": {
            "get": {
                "description": "Get One users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "FetchOneUserById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "example:257d3552-c186-4c23-aa5d-1ea53f453e2a",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/constants.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "constants.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "Invalid email format"
                }
            }
        }
    }
}